When the vtkChartsQtVega is built js files are copied in the build dir.
Now JavaScript modules and source files, needed for the generation
of the Vega scene graph, are placed respectively under the
Charts/QtVega/jsmodules and the Charts/QtVega/jslib paths of the
source folder.

Moreover when the vtkChartsQtVega module is built they are
automatically copied to the same relative paths into the build folder.

Change-Id: I2b74d405e26e30ced7ed6c43921249f7d2882c7f