Implemented `vtkSVGPath` and related classes.
vtkSVGPath is an object that allow the user to defines an SVG path
as a sequence of commands where each command is an instance
of a class derived from vtkSVGPathCommand. Moreover it is able
to compute a collection of arrays of 2D points. Each array contains
the vertices of a polyline which approximates a sub-path defined
into the path object.

For each SVG path command a class has been implemented.
All these classes derive from vtkSVGPathCommand. This is an
abstract class which represents a generic SVG path command.
It provides properties and behaviors common to all SVG path
command classes.

The vtkSVGPathContext is a data structure used by vtkSVGPath
 methods for keeping track of path information while processing
SVG path commands.

A test unit providing basic testing for vtkSVGPath has been
implemented.

Change-Id: I76e9e89e8b24b36520d4762b8c0d7db1e2632cac