A modification to d3.js module has been needed for fixing an issue.
The problem is that the Qt JavaScript engine evaluates the following
expression:

x = d + (d = t);
as:
d = t;  x = d + d;
while the correct way is:
d0 = d;  d = t;  x = d0 + d;

This problem affected the evaluation of contol points in the
pathCache array, leading to wrong path rendering when cubic
`curveTo` command are used.

In order to solve this issue I needed to modify manually the source
code of the `d3` module.

Change-Id: I9c8c769ab337f3078956217d2b22a52901978de5