Now axes are correctly placed according to the layer property value.
The layer property for axes is not removed anymore from the scene
DAG. The layer property for axes is needed for the correct
displacement of each axis at the front or at the back respect with
mark items which made up the chart.

The UpdateGroupItem of the vtkVegaRenderer class has been
modified in order to keep into account the value of the layer
property and to draw correctly each axes at front or at back.

Change-Id: I421105f4a52c86079e66c5516cf919275aebdc3a